Move compareResourceCommands into resourceDisplay
Consolidate the resource command comparator alongside the other resource display helpers and remove the standalone resourceCommands.ts module.
